apps/saip_be/requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
django-rest-knox~=4.2.0
django-cors-headers~=3.13.0
django_extensions~=3.2.1
gunicorn~=20.1.0
psycopg2-binary~=2.9.5
django-object-actions~=4.1.0
daphne~=4.0.0
channels~=4.0.0